chore(refactor): consolidate recipe detail templates into single view
