+++ src/routes/recipes/detail.rs
@@ -3,13 +3,18 @@ use axum::{
     response::{IntoResponse, Redirect},
 };
 
-use imkitchen_recipe::user::UserView;
+use evento::cursor::{Args, ReadResult};
+use imkitchen_recipe::{
+    rating,
+    user::{RecipesQuery, UserView, UserViewList},
+    user_stat::UserStatView,
+};
 use imkitchen_shared::recipe::{IngredientUnitFormat, RecipeType};
 
 use crate::{
     auth::AuthUser,
     routes::AppState,
-    template::{ForbiddenTemplate, Status, Template, filters},
+    template::{NotFoundTemplate, Status, Template, filters},
 };
 
 #[derive(askama::Template)]
@@ -23,8 +28,8 @@ pub struct DeleteModalTemplate {
 }
 
 #[derive(askama::Template)]
-#[template(path = "partials/recipes-share-button.html")]
-pub struct ShareButtonTemplate<'a> {
+#[template(path = "partials/recipes-detail-share-button.html")]
+pub struct CommunityDetailShareButtonTemplate<'a> {
     pub id: &'a str,
     pub is_shared: bool,
 }
@@ -42,6 +47,10 @@ pub struct DetailTemplate {
     pub current_path: String,
     pub user: AuthUser,
     pub recipe: UserView,
+    pub stat: UserStatView,
+    pub rating: rating::Rating,
+    pub cook_recipes: ReadResult<UserViewList>,
+    pub similar_recipes: ReadResult<UserViewList>,
 }
 
 impl Default for DetailTemplate {
@@ -50,6 +59,10 @@ impl Default for DetailTemplate {
             current_path: "recipes".to_owned(),
             user: AuthUser::default(),
             recipe: Default::default(),
+            stat: UserStatView::default(),
+            cook_recipes: Default::default(),
+            similar_recipes: Default::default(),
+            rating: Default::default(),
         }
     }
 }
@@ -63,14 +76,172 @@ pub async fn page(
 ) -> impl IntoResponse {
     let recipe = crate::try_page_response!(opt: app.recipe_query.user(&id), template);
 
-    if recipe.owner_id != user.id {
-        return template.render(ForbiddenTemplate).into_response();
+    if recipe.owner_id != user.id && !recipe.is_shared {
+        return template.render(NotFoundTemplate).into_response();
+    }
+
+    let stat =
+        crate::try_page_response!(app.recipe_query.find_user_stat(&recipe.owner_id), template)
+            .unwrap_or_default();
+
+    let rating = crate::try_page_response!(
+        app.recipe_cmd.rating.load(&recipe.id, &recipe.owner_id),
+        template
+    )
+    .to_owned();
+
+    let exclude_ids = vec![recipe.id.to_owned()];
+
+    let cook_recipes = crate::try_page_response!(
+        app.recipe_query.filter_user(RecipesQuery {
+            exclude_ids: Some(exclude_ids),
+            user_id: Some(recipe.owner_id.to_owned()),
+            recipe_type: None,
+            cuisine_type: None,
+            is_shared: Some(true),
+            dietary_restrictions: vec![],
+            dietary_where_any: false,
+            sort_by: imkitchen_recipe::user::SortBy::RecentlyAdded,
+            args: Args::forward(2, None),
+        }),
+        template
+    );
+
+    let mut exclude_ids = cook_recipes
+        .edges
+        .iter()
+        .map(|n| n.node.id.to_owned())
+        .collect::<Vec<_>>();
+
+    exclude_ids.push(recipe.id.to_owned());
+
+    let mut similar_recipes = crate::try_page_response!(
+        app.recipe_query.filter_user(RecipesQuery {
+            exclude_ids: Some(exclude_ids.to_vec()),
+            user_id: None,
+            recipe_type: Some(recipe.recipe_type.0.to_owned()),
+            cuisine_type: Some(recipe.cuisine_type.0.to_owned()),
+            is_shared: Some(true),
+            dietary_restrictions: recipe.dietary_restrictions.0.to_vec(),
+            dietary_where_any: false,
+            sort_by: imkitchen_recipe::user::SortBy::RecentlyAdded,
+            args: Args::forward(6, None),
+        }),
+        template
+    );
+
+    if similar_recipes.edges.len() < 6 {
+        let mut similar_ids = similar_recipes
+            .edges
+            .iter()
+            .map(|n| n.node.id.to_owned())
+            .collect::<Vec<_>>();
+        similar_ids.extend(exclude_ids.to_vec());
+
+        let more_recipes = crate::try_page_response!(
+            app.recipe_query.filter_user(RecipesQuery {
+                exclude_ids: Some(similar_ids),
+                user_id: None,
+                recipe_type: Some(recipe.recipe_type.0.to_owned()),
+                cuisine_type: Some(recipe.cuisine_type.0.to_owned()),
+                is_shared: Some(true),
+                dietary_restrictions: recipe.dietary_restrictions.0.to_vec(),
+                dietary_where_any: true,
+                sort_by: imkitchen_recipe::user::SortBy::RecentlyAdded,
+                args: Args::forward(6, None),
+            }),
+            template
+        );
+
+        similar_recipes.edges.extend(more_recipes.edges);
+    }
+
+    if similar_recipes.edges.len() < 6 {
+        let mut similar_ids = similar_recipes
+            .edges
+            .iter()
+            .map(|n| n.node.id.to_owned())
+            .collect::<Vec<_>>();
+        similar_ids.extend(exclude_ids.to_vec());
+
+        let more_recipes = crate::try_page_response!(
+            app.recipe_query.filter_user(RecipesQuery {
+                exclude_ids: Some(similar_ids),
+                user_id: None,
+                recipe_type: Some(recipe.recipe_type.0.to_owned()),
+                cuisine_type: None,
+                is_shared: Some(true),
+                dietary_restrictions: recipe.dietary_restrictions.0.to_vec(),
+                dietary_where_any: false,
+                sort_by: imkitchen_recipe::user::SortBy::RecentlyAdded,
+                args: Args::forward(6, None),
+            }),
+            template
+        );
+
+        similar_recipes.edges.extend(more_recipes.edges);
+    }
+
+    if similar_recipes.edges.len() < 6 {
+        let mut similar_ids = similar_recipes
+            .edges
+            .iter()
+            .map(|n| n.node.id.to_owned())
+            .collect::<Vec<_>>();
+        similar_ids.extend(exclude_ids.to_vec());
+
+        let more_recipes = crate::try_page_response!(
+            app.recipe_query.filter_user(RecipesQuery {
+                exclude_ids: Some(similar_ids),
+                user_id: None,
+                recipe_type: Some(recipe.recipe_type.0.to_owned()),
+                cuisine_type: None,
+                is_shared: Some(true),
+                dietary_restrictions: recipe.dietary_restrictions.0.to_vec(),
+                dietary_where_any: true,
+                sort_by: imkitchen_recipe::user::SortBy::RecentlyAdded,
+                args: Args::forward(6, None),
+            }),
+            template
+        );
+
+        similar_recipes.edges.extend(more_recipes.edges);
+    }
+
+    if similar_recipes.edges.len() < 6 {
+        let mut similar_ids = similar_recipes
+            .edges
+            .iter()
+            .map(|n| n.node.id.to_owned())
+            .collect::<Vec<_>>();
+        similar_ids.extend(exclude_ids);
+
+        let more_recipes = crate::try_page_response!(
+            app.recipe_query.filter_user(RecipesQuery {
+                exclude_ids: Some(similar_ids),
+                user_id: None,
+                recipe_type: Some(recipe.recipe_type.0.to_owned()),
+                cuisine_type: None,
+                is_shared: Some(true),
+                dietary_restrictions: vec![],
+                dietary_where_any: false,
+                sort_by: imkitchen_recipe::user::SortBy::RecentlyAdded,
+                args: Args::forward(6, None),
+            }),
+            template
+        );
+
+        similar_recipes.edges.extend(more_recipes.edges);
     }
 
     template
         .render(DetailTemplate {
             user,
             recipe,
+            stat,
+            cook_recipes,
+            similar_recipes,
+            rating,
             ..Default::default()
         })
         .into_response()
@@ -97,7 +268,7 @@ pub async fn share_to_community_action(
     );
 
     template
-        .render(ShareButtonTemplate {
+        .render(CommunityDetailShareButtonTemplate {
             id: &id,
             is_shared: true,
         })
@@ -114,7 +285,7 @@ pub async fn make_private_action(
     crate::try_response!(app.recipe_cmd.make_private(&id, &user.id), template);
 
     template
-        .render(ShareButtonTemplate {
+        .render(CommunityDetailShareButtonTemplate {
             id: &id,
             is_shared: false,
         })
@@ -165,4 +336,114 @@ pub async fn delete_status(
 
 pub async fn delete_modal(template: Template, Path((id,)): Path<(String,)>) -> impl IntoResponse {
     template.render(DeleteModalTemplate { id })
+}
+
+pub async fn check_in(
+    template: Template,
+    user: AuthUser,
+    State(app): State<AppState>,
+    Path((id,)): Path<(String,)>,
+) -> impl IntoResponse {
+    crate::try_response!(app.recipe_cmd.rating.view(&id, &user.id), template);
+    "<div></div>".into_response()
+}
+
+#[derive(askama::Template)]
+#[template(path = "partials/recipes-detail-like-button.html")]
+pub struct LikeButtonTemplate {
+    pub id: String,
+    pub total: u64,
+    pub liked: bool,
+    pub unliked: bool,
+}
+
+pub async fn check_like(
+    template: Template,
+    user: AuthUser,
+    State(app): State<AppState>,
+    Path((id,)): Path<(String,)>,
+) -> impl IntoResponse {
+    crate::try_response!(app.recipe_cmd.rating.check_like(&id, &user.id), template);
+
+    let recipe = crate::try_response!(anyhow_opt: app.recipe_query.user(&id), template);
+
+    (
+        [("ts-swap", "skip")],
+        template.render(LikeButtonTemplate {
+            id,
+            total: recipe.total_ulikes(),
+            liked: true,
+            unliked: false,
+        }),
+    )
+        .into_response()
+}
+
+pub async fn uncheck_like(
+    template: Template,
+    user: AuthUser,
+    State(app): State<AppState>,
+    Path((id,)): Path<(String,)>,
+) -> impl IntoResponse {
+    crate::try_response!(app.recipe_cmd.rating.uncheck_like(&id, &user.id), template);
+
+    let recipe = crate::try_response!(anyhow_opt: app.recipe_query.user(&id), template);
+
+    (
+        [("ts-swap", "skip")],
+        template.render(LikeButtonTemplate {
+            id,
+            total: recipe.total_ulikes(),
+            liked: false,
+            unliked: false,
+        }),
+    )
+        .into_response()
+}
+
+pub async fn check_unlike(
+    template: Template,
+    user: AuthUser,
+    State(app): State<AppState>,
+    Path((id,)): Path<(String,)>,
+) -> impl IntoResponse {
+    crate::try_response!(app.recipe_cmd.rating.check_unlike(&id, &user.id), template);
+
+    let recipe = crate::try_response!(anyhow_opt: app.recipe_query.user(&id), template);
+
+    (
+        [("ts-swap", "skip")],
+        template.render(LikeButtonTemplate {
+            id,
+            total: recipe.total_ulikes(),
+            liked: false,
+            unliked: true,
+        }),
+    )
+        .into_response()
+}
+
+pub async fn uncheck_unlike(
+    template: Template,
+    user: AuthUser,
+    State(app): State<AppState>,
+    Path((id,)): Path<(String,)>,
+) -> impl IntoResponse {
+    crate::try_response!(
+        app.recipe_cmd.rating.uncheck_unlike(&id, &user.id),
+        template
+    );
+
+    let recipe = crate::try_response!(anyhow_opt: app.recipe_query.user(&id), template);
+
+    (
+        [("ts-swap", "skip")],
+        template.render(LikeButtonTemplate {
+            id,
+            total: recipe.total_ulikes(),
+            liked: false,
+            unliked: false,
+        }),
+    )
+        .into_response()
 }
