+++ src/routes/recipes/mod.rs
@@ -1,5 +1,4 @@
 pub mod community;
-pub mod community_detail;
 pub mod detail;
 pub mod edit;
 pub mod import;
