+++ templates/partials/recipes-delete-modal.html
@@ -23,7 +23,7 @@
         class="flex-1 px-4 py-2 bg-gray-200 text-gray-700 font-semibold rounded-lg hover:bg-gray-300">
         {{ "Cancel"|t }}
       </button>
-      <button ts-trigger="click" ts-req="/recipes/{{ id }}/detail/delete" ts-req-method="POST"
+      <button ts-trigger="click" ts-req="/recipes/{{ id }}/delete" ts-req-method="POST"
         class="flex-1 px-4 py-2 bg-red-600 text-white font-semibold rounded-lg hover:bg-red-700 cursor-pointer">
         {{ "Delete"|t }}
       </button>
