+++ templates/partials/recipes-detail-share-button.html
@@ -1,0 +1,21 @@
+{% if is_shared %}
+<button ts-req="/recipes/{{ id }}/make-private"
+  class="flex items-center gap-2 px-4 py-2 bg-gray-100 hover:bg-gray-200 rounded-full transition">
+  <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
+      d="M12 15v2m-6 4h12a2 2 0 002-2v-6a2 2 0 00-2-2H6a2 2 0 00-2 2v6a2 2 0 002 2zm10-10V7a4 4 0 00-8 0v4h8z">
+    </path>
+  </svg>
+  <span class="hidden sm:inline font-semibold">{{ "Make Private"|t }}</span>
+</button>
+{% else %}
+<button ts-req="/recipes/{{ id }}/share-to-community"
+  class="flex items-center gap-2 px-4 py-2 bg-gray-100 hover:bg-gray-200 rounded-full transition">
+  <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
+      d="M3.055 11H5a2 2 0 012 2v1a2 2 0 002 2 2 2 0 012 2v2.945M8 3.935V5.5A2.5 2.5 0 0010.5 8h.5a2 2 0 012 2 2 2 0 104 0 2 2 0 012-2h1.064M15 20.488V18a2 2 0 012-2h3.064M21 12a9 9 0 11-18 0 9 9 0 0118 0z">
+    </path>
+  </svg>
+  <span class="hidden sm:inline font-semibold">{{ "Make Public"|t }}</span>
+</button>
+{% endif %}
