+++ templates/recipes-community.html
@@ -81,7 +81,7 @@
     </div>
     {% else %}
     {% for recipe in recipes.edges %}
-    <a id="recipe-{{ recipe.node.id }}" href="/recipes/{{ recipe.node.id }}/community" {% if let (true, true,
+    <a id="recipe-{{ recipe.node.id }}" href="/recipes/{{ recipe.node.id }}" {% if let (true, true,
       Some(cursor))=(loop.last, recipes.page_info.has_next_page, recipes.page_info.end_cursor.to_owned()) %}
       ts-req="/recipes/community?after={{ cursor.to_string() }}{% if let Some(recipe_type) = query.recipe_type %}&recipe_type={{ recipe_type }}{% endif %}{% if let Some(cuisine_type) = query.cuisine_type %}&cuisine_type={{ cuisine_type }}{% endif %}{% for restriction in query.dietary_restrictions %}&dietary_restrictions={{ restriction }}{% endfor %}"
       ts-req-method="GET" ts-req-selector="children #recipes-list" ts-swap="afterend" ts-trigger="visible once" {% endif
